added an RNAseq supackage
Added a subpackage for RNAseq related stuff. Added a script that automates the initial processing steps for going from raw sequence read data to aligned and sorted feature counts. Works so far with dummy stand-in programs (hisat2, samtools, featureCounts) and dummy data, but currently testing on xucube with the real programs and data.